Revert "Revert "First step of spliting home component (#145)" (#146)"
This reverts commit d7f37098907e9a9075458e9119abdf2cb400cc40.
